(WIP) Rename things better, start on specs
